requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.22.0
opencv-python>=4.6.0
matplotlib>=3.5.0
pyyaml>=6.0
scikit-learn>=1.0.0
scikit-image>=0.19.0
tensorboard>=2.10.0
tqdm>=4.64.0
lpips>=0.1.4
timm>=0.6.12  # For advanced model architectures
pycocotools>=2.0.6  # For object detection evaluation